LayoutTests/http/tests/in-app-browser-privacy/context-string-preflight-expected.txt

CONSOLE MESSAGE: Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/status-404-without-body.py due to access control checks.
CONSOLE MESSAGE: Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/status-404-without-body.py due to access control checks.
Tests that not successful preflight responses make preflight failing

PASS: NetworkError:  A network error occurred.
PASS Context string for CORS preflight correctly captured.
